+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Add getConversation() method to Toot class
 * Log suspended accounts
 
 ### v0.48.10